Sprite offsets
From this commit Sprites are able to have an offset, in fact
they can have pictures that will only be played at the start or
if the sprite is set to use an earlier image using currentSprite
Variable.

In the need of this the Character constructor has been refitted
with respecting parameters, which may need further implementation.

Also the background images of the menu have now an offset of 0 as
they are also Sprites.

The default JSONs of the monster and ninja now also have an
value for the walk- and jump- offset set to 0 which now needs to
be adjusted.

*complete Javadoc descriptions
*detect more wrong parameters early at setCurrentPicture in Sprites

respecting Issues:
#2